requirements.txt
langchain-openai == 0.1.8
python-dotenv == 1.0.1
langchain == 0.2.1
langchain-community == 0.2.1
